// src/modules/orders/enums/order-status.enum.ts
export enum OrderStatus {
    PENDING = 'PENDING',
    SEARCHING = 'SEARCHING',
    ACCEPTED = 'ACCEPTED',
    PROVIDER_ON_THE_WAY = 'PROVIDER_ON_THE_WAY',
    PROVIDER_ARRIVED = 'PROVIDER_ARRIVED',
    IN_PROGRESS = 'IN_PROGRESS',
    COMPLETED = 'COMPLETED',
    CANCELLED = 'CANCELLED',
}

export enum OfferStatus {
    PENDING = 'PENDING',
    ACCEPTED = 'ACCEPTED',
    REJECTED = 'REJECTED',
    CANCELLED = 'CANCELLED',
}